Fix wrong date
Previously the JSON data from ICA only contained a date without a year.
Before this commit, the current year was added to make it a complete
year. This does not work when some shipping slot entries fall within
the current year and some entries in the next year. Fortunately, a new
field in the JSON data has been added which contains the complete date.